+++ vibetags-annotations/src/main/java/se/deversity/vibetags/annotations/AIKeepInSync.java
@@ -27,7 +27,7 @@ import java.lang.annotation.Target;
  * }</pre>
  */
 @Retention(RetentionPolicy.SOURCE)
-@Target({ElementType.TYPE, ElementType.METHOD, ElementType.FIELD})
+@Target({ElementType.TYPE, ElementType.METHOD, ElementType.CONSTRUCTOR, ElementType.FIELD})
 public @interface AIKeepInSync {
 
     /**