+++ vibetags-annotations/src/main/java/se/deversity/vibetags/annotations/AIThreadAffinity.java
@@ -28,7 +28,7 @@ import java.lang.annotation.Target;
  * }</pre>
  */
 @Retention(RetentionPolicy.SOURCE)
-@Target({ElementType.TYPE, ElementType.METHOD})
+@Target({ElementType.TYPE, ElementType.METHOD, ElementType.CONSTRUCTOR})
 public @interface AIThreadAffinity {
 
     /** Which thread an element is pinned to. */